requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
mysql-connector-python==9.0.0
pyOpenSSL==24.2.1
requests==2.32.3
schedule